Restructure service bundle, fix names
