+++ edu.kit.ipd.sdq.kamp.ruledsl/META-INF/MANIFEST.MF
@@ -26,7 +26,8 @@ Require-Bundle: org.eclipse.xtext,
  org.eclipse.pde;bundle-version="3.12.3",
  org.eclipse.pde.core;bundle-version="3.11.1",
  tools.vitruv.framework.util;bundle-version="0.2.0",
- org.eclipse.xtext.ui;bundle-version="2.11.0"
+ org.eclipse.xtext.ui;bundle-version="2.11.0",
+ edu.kit.ipd.sdq.kamp4bp
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: edu.kit.ipd.sdq.kamp.ruledsl,
  edu.kit.ipd.sdq.kamp.ruledsl.generator,
