+++ edu.kit.ipd.sdq.kamp.ruledsl/src/edu/kit/ipd/sdq/kamp/ruledsl/KampRuleLanguage.xtext
@@ -9,7 +9,7 @@ import "http://www.kit.edu/ipd/sdq/vitruvius/dsls/mirbase/MirBase" as mirbase
 RuleFile:
 	MirBaseFile
 //	'plugin-name' ':' name=ValidID
-	(rules+=KampRule)+
+	(rules+=KampRule)*
 ;
 
 KampRule:
