+++ edu.kit.ipd.sdq.kamp.ruledsl/src/edu/kit/ipd/sdq/kamp/ruledsl/service/IRule.java
@@ -1,5 +1,7 @@
 package edu.kit.ipd.sdq.kamp.ruledsl.service;
 
+import edu.kit.ipd.sdq.kamp4bp.core.BPArchitectureVersion;
+
 public interface IRule {
-	void apply();
+	void apply(BPArchitectureVersion version);
 }
